Move firebase initialization to own file and delete unnecessary service worker
