+++ src/index.tsx
@@ -3,22 +3,13 @@ import ReactDOM from 'react-dom';
 import './index.scss';
 
 import { Provider } from 'react-redux';
-import firebase from 'firebase/app';
-import 'firebase/auth';
-import 'firebase/firestore';
+
 import { createStore, combineReducers} from 'redux';
 import {ReactReduxFirebaseProvider, firebaseReducer} from 'react-redux-firebase';
 import { createFirestoreInstance, firestoreReducer } from 'redux-firestore';
 
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.REACT_APP_FIREBASE_DATABASE_URL,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID
-};
+import firebase from './firebase';
+
 
 // react-redux-firebase config
 const rrfConfig = {
@@ -26,11 +17,6 @@ const rrfConfig = {
   useFirestoreForProfile: true // Firestore for Profile instead of Realtime DB
 };
 
-// Initialize firebase instance
-firebase.initializeApp(firebaseConfig);
-
-// Initialize other services on firebase instance
-firebase.firestore();
 
 // Add firebase to reducers
 const rootReducer = combineReducers({
